An attempt to add stuff to script/bootstrap for new installs.
Ideally the goal is you clone dotfiles & run script/bootstrap, then
everything works.

At least the script/bootstrap should keep fairly up to date with changes
outside of the main Rakefile, even if it isn't tested.

This mainly adds the setup for base16 theme - which finally fixes the
problem of unreadable symlinks etc in shell (or hideous green in vim...)