Separate ZEUSSOCK per-project for multiple zeuses. (Set every `cd`)
Moving the ZEUSSOCK outside the project root is necessary when the
projects are mounted in a FUSE filesystem, as is the case when the
folder is ENCFS protected (or so i'm told with vagrant).

However setting it as an ENV constant means you can only have one zeus
process running, and only one app.  Dynamically changing it means we can
run multiple rails projects under zeus simultaneously.  This is handy
when developing SOA apps.